requirements.txt
Flask==3.0.3
Requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
Jinja2==3.1.4
